WebDriver: unable to run tests using debug safaridriver, hitting ASSERT(!decoder.destinationID()) https://bugs.webkit.org/show_bug.cgi?id=282055
Reviewed by NOBODY (OOPS!).

We hit this assertion when dispatching WebAutomationSessionProxy::Messages::EvaluateJavaScriptFunction.
This code was refactored on the UIProcess-side to call sendToProcessContainingFrame(), which includes
a destinationID. However, a destinationID is not expected because WebAutomationSessionProxy is a singleton
in WebProcess which outlives specific pages or frames.

To work around this, introduce a new sendToProcessContainingFrame() funnel method which does not include
the destinationID when dispatching the two destinationID-less messages to WebAutomationSessionProxy.

* Source/WebKit/UIProcess/Automation/WebAutomationSession.cpp:
(WebKit::WebAutomationSession::evaluateJavaScriptFunction):
(WebKit::WebAutomationSession::takeScreenshot):

* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::sendToProcessContainingFrameWithoutDestinationIdentifier): Added.
Adjust template instantiation macros so that the affected WebAutomationSessionProxy messages are sent
without a destinationId.